requirements.txt
asgiref==3.9.1
Django==5.2.3
django-js-asset==3.1.2
django-mptt==0.17.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==10.2.0
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
